requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1

flask-restful
mysql-connector-python
psycopg2-binary
passlib
Flask-JWT-Extended
email-validator
requests==2.25.0